tools/mcp-skill-router/requirements.txt
mcp
chromadb
openai
pyyaml
